--- tests/test_000_imports.py
@@ -1,4 +1,0 @@
-import pytest
-
-def test_underworld3_modules():
-    import underworld3 as uw
