+++ tests/test_0_imports.py
@@ -1,0 +1,4 @@
+import pytest
+
+def test_underworld3_modules():
+    import underworld3 as uw
